Trade Zone Wealth Remapping
UniqueLayouts' mapping is working

A check is done to make sure it tries to find the stat enough times
at sheets layed out differently.

The current Command's index is public now.

The alerts will show the line at which the error occurs.

Small fix that made the sheet restriction search in 0 sheets,
instead the 1 sheet it should.