Library removed.
Library was unable to deal with the way we have set up sheets.
Going the manual way not

This does not compile. This is merely committed to have access to this
other places.

References to the Library is removed, but the code has not been refactored to
take this into consideration.