remove 47 types of rogue pixels
Rogue pixels here defined as pixels colors that are adjescent but
not equal to that of a nation. The rogue pixel type
with most rogue pixels, was one with 8. All other rogue pixel types
had less than this. They were deleted and replaced with transparent.

Note,
NewRabat has some of its size from a little offshoot in mainland Valublid