getStatType now takes the entire selection string
This is to allow for more specific and versitile specifications
of which stats are what types, even inside complex objects.

It is necessary to allow for objects to have types too, as types
are now not only used for displaying a symbol in
the corner of a table cell, but also
controls which stats are considered protected.